clean up and remove unused imports
